// src/api/generated/models/addressUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */

/**
 * Partially updates an address record.
 */
export interface AddressUpdate {
  street?: string | null
  house_number?: string | null
  zip_code?: string | null
  city?: string | null
  latitude?: number | null
  longitude?: number | null
}