// src/api/generated/models/officeCreate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */
import type { AddressCreate } from './addressCreate'
import type { OpeningHours } from './openingHours'

/**
 * Validate the payload used to create an office.
 */
export interface OfficeCreate {
  /**
   * @minLength 3
   * @maxLength 150
   */
  name: string
  description?: string | null
  contact_email?: string | null
  phone?: string | null
  /** @maxItems 50 */
  services?: string[]
  opening_hours?: OpeningHours | null
  address?: AddressCreate | null
}