// src/api/generated/models/officeUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */
import type { AddressUpdate } from './addressUpdate'
import type { OpeningHours } from './openingHours'

/**
 * Validate a partial office update payload.
 */
export interface OfficeUpdate {
  name?: string | null
  description?: string | null
  contact_email?: string | null
  phone?: string | null
  services?: string[] | null
  opening_hours?: OpeningHours | null
  address?: AddressUpdate | null
  /**
   * @minLength 3
   * @maxLength 500
   */
  change_reason: string
}